--- a/src/postgres/inventory/paginator/PgPaginatorOrderingAttributes.ts
+++ b/src/postgres/inventory/paginator/PgPaginatorOrderingAttributes.ts
@@ -66,12 +66,12 @@ implements Paginator.Ordering<TInput, PgClassType.Value, AttributesCursor> {
       throw new Error('Before cursor must be a value tuple of the correct length.')
 
     const aliasIdentifier = Symbol()
-    const aliasIdentifier2 = Symbol()
+    const cteIdentifier = Symbol()
     const fromSql = this.pgPaginator.getFromEntrySql(input)
     const conditionSql = this.pgPaginator.getConditionSql(input)
 
     const query = sql.compile(sql.query`
-      with ${sql.identifier(aliasIdentifier2)} as (
+      with ${sql.identifier(cteIdentifier)} as (
         -- The standard select/from clauses up top.
         select ${sql.identifier(aliasIdentifier)}.*
         from ${fromSql} as ${sql.identifier(aliasIdentifier)}
@@ -99,7 +99,7 @@ implements Paginator.Ordering<TInput, PgClassType.Value, AttributesCursor> {
 
         -- If we have an offset, add that as well.
         ${_offset != null ? sql.query`offset ${sql.value(_offset)}` : sql.query``}
-      ) select to_json(${sql.identifier(aliasIdentifier2)}) as value from ${sql.identifier(aliasIdentifier2)};
+      ) select to_json(${sql.identifier(cteIdentifier)}) as value from ${sql.identifier(cteIdentifier)};
     `)
 
     let { rows } = await client.query(query)