--- a/src/postgres/inventory/paginator/PgPaginatorOrderingOffset.ts
+++ b/src/postgres/inventory/paginator/PgPaginatorOrderingOffset.ts
@@ -121,20 +121,20 @@ implements Paginator.Ordering<TInput, TItemValue, OffsetCursor> {
     }
 
     const aliasIdentifier = Symbol()
-    const aliasIdentifier2 = Symbol()
+    const cteIdentifier = Symbol()
     const fromSql = this.pgPaginator.getFromEntrySql(input)
     const conditionSql = this.pgPaginator.getConditionSql(input)
 
     // Construct our Sql query that will actually do the selecting.
     const query = sql.compile(sql.query`
-      with ${sql.identifier(aliasIdentifier2)} as (
+      with ${sql.identifier(cteIdentifier)} as (
         select ${sql.identifier(aliasIdentifier)}.*
         from ${fromSql} as ${sql.identifier(aliasIdentifier)}
         where ${conditionSql}
         ${this.orderBy ? sql.query`order by ${this.orderBy}` : sql.query``}
         offset ${sql.value(offset)}
         limit ${limit != null ? sql.value(limit) : sql.query`all`}
-      ) select to_json(${sql.identifier(aliasIdentifier2)}) as value from ${sql.identifier(aliasIdentifier2)};
+      ) select to_json(${sql.identifier(cteIdentifier)}) as value from ${sql.identifier(cteIdentifier)};
     `)
 
     // Send our query to Postgres.