--- a/src/postgres/inventory/paginator/PgPaginatorOrderingOffset.ts
+++ b/src/postgres/inventory/paginator/PgPaginatorOrderingOffset.ts
@@ -131,13 +131,13 @@ implements Paginator.Ordering<TInput, TItemValue, OffsetCursor> {
       -- applied. This ensures the to_json function is only called on the final
       -- results of the query.
       with ${sql.identifier(cteIdentifier)} as (
-        select ${sql.identifier(aliasIdentifier)}.*
+        select ${sql.identifier(aliasIdentifier)} as value
         from ${fromSql} as ${sql.identifier(aliasIdentifier)}
         where ${conditionSql}
         ${this.orderBy ? sql.query`order by ${this.orderBy}` : sql.query``}
         offset ${sql.value(offset)}
         limit ${limit != null ? sql.value(limit) : sql.query`all`}
-      ) select to_json(${sql.identifier(cteIdentifier)}) as value from ${sql.identifier(cteIdentifier)};
+      ) select to_json(${sql.identifier(cteIdentifier)}.value) as value from ${sql.identifier(cteIdentifier)};
     `)
 
     // Send our query to Postgres.