This uses Postgres common table expressions to make sure the `to_json`-function is only called on the results of a pagination, as discussed in #380

---

@Benjie's explanation for this PR:

It seems that PostgreSQL sometimes performs the `to_json()` whilst ordering the rows and before applying the limit - this means that a `select to_json(tbl) from tbl order by col limit 10` is **significantly** slower than both `select tbl from tbl order by col limit 10` and `select to_json(tbl) from tbl limit 10`. To work around this we use a common table expression (CTE) as an optimisation barrier to first find the matching rows, and only after doing so perform to_json on the relevant rows (and not the irrelevant ones). This yields a significant performance improvement.