@benjie 2017-03-16
I have an issue with `intSetQuery` on #342 also.