@benjie 2017-05-24
This is still failing the tests on `intSetQuery` - I have the same issue on #342 where I am making heavy use of CTEs. I believe the issue is that each row is returned as an object with one property, the value for which is an integer. I believe if you were to detect this case (where the function returns a setof scalars) you could handle it by mapping `const rows = rawRows.map(row => Object.values(row)[0])`.