@danielbuechele 2017-07-06
I added the requested changes and removed the usage of `.*` which seemed to be a performance issue in some cases.