@benjie 2017-07-06
Does this give a quantifiable performance improvement? My tests previously suggested it wasn't just the to_json that caused the issue, but also the referencing the table without the `.*` (which seemed to be processed before the filter rather than after, as is the case when an index is present):

https://gist.github.com/benjie/02df5fdbd7ee8315712fff83007f3cd1 explain.sql; i.e.:

with `.*`:

```
[test] # explain analyse select sketch_integration_event_run_step.* from sketch_integration_event_run_step order by type limit 10;
  | QUERY PLAN
  | ---------------------------------------------------------------------------------------------------------------------------------------------------------
  | Limit  (cost=79913.03..79913.06 rows=10 width=807) (actual time=556.499..556.500 rows=10 loops=1)
  | ->  Sort  (cost=79913.03..82010.18 rows=838859 width=807) (actual time=556.497..556.498 rows=10 loops=1)
  | Sort Key: type
  | Sort Method: top-N heapsort  Memory: 26kB
  | ->  Seq Scan on sketch_integration_event_run_step  (cost=0.00..61785.59 rows=838859 width=807) (actual time=0.019..219.488 rows=838708 loops=1)
  | Planning time: 0.084 ms
  | Execution time: 556.536 ms
  | (7 rows)
  |  
  | Time: 556.919 ms
```

without `.*`:

```
  | [test] # explain analyse select sketch_integration_event_run_step from sketch_integration_event_run_step order by type limit 10;
  | QUERY PLAN
  | ----------------------------------------------------------------------------------------------------------------------------------------------------------
  | Limit  (cost=79913.03..79913.06 rows=10 width=837) (actual time=7939.256..7939.259 rows=10 loops=1)
  | ->  Sort  (cost=79913.03..82010.18 rows=838859 width=837) (actual time=7939.255..7939.258 rows=10 loops=1)
  | Sort Key: type
  | Sort Method: top-N heapsort  Memory: 26kB
  | ->  Seq Scan on sketch_integration_event_run_step  (cost=0.00..61785.59 rows=838859 width=837) (actual time=0.033..7150.607 rows=838708 loops=1)
  | Planning time: 0.089 ms
  | Execution time: 7939.289 ms
  | (7 rows)
  |  
  | Time: 7939.653 ms
```