@danielbuechele 2017-07-07
In my case (using row level security) the `.*` operator doesn't make any difference. However using CTEs over not using them makes a big difference.

**no `.*`, no CTE**
```
Limit  (cost=31989.50..31989.63 rows=50 width=62) (actual time=3917.921..3917.937 rows=50 loops=1)
  ->  Sort  (cost=31989.50..32277.03 rows=115009 width=62) (actual time=3917.920..3917.934 rows=50 loops=1)
        Sort Key: __local_0__.id
        Sort Method: top-N heapsort  Memory: 161kB
        ->  Subquery Scan on __local_0__  (cost=0.00..28168.99 rows=115009 width=62) (actual time=0.343..3839.865 rows=110587 loops=1)
              ->  Seq Scan on articles __local_0___1  (cost=0.00..26731.38 rows=115009 width=1930) (actual time=0.262..221.573 rows=110587 loops=1)
                    Filter: (status = 'PUBLISHED'::status)
                    Rows Removed by Filter: 6208
Planning time: 2.721 ms
Execution time: 3917.965 ms
```

**with `.*`, no CTE**
```
Limit  (cost=31989.50..31989.63 rows=50 width=62) (actual time=4030.565..4030.584 rows=50 loops=1)
  ->  Sort  (cost=31989.50..32277.03 rows=115009 width=62) (actual time=4030.563..4030.579 rows=50 loops=1)
        Sort Key: __local_0__.id
        Sort Method: top-N heapsort  Memory: 161kB
        ->  Subquery Scan on __local_0__  (cost=0.00..28168.99 rows=115009 width=62) (actual time=0.451..3947.324 rows=110587 loops=1)
              ->  Seq Scan on articles __local_0___1  (cost=0.00..26731.38 rows=115009 width=1930) (actual time=0.374..244.449 rows=110587 loops=1)
                    Filter: (status = 'PUBLISHED'::status)
                    Rows Removed by Filter: 6208
Planning time: 3.815 ms
Execution time: 4030.613 ms
```

**no `.*`, with CTE**
```
CTE Scan on test  (cost=31702.11..31703.23 rows=50 width=32) (actual time=201.214..203.636 rows=50 loops=1)
  CTE test
    ->  Limit  (cost=31701.98..31702.11 rows=50 width=1876) (actual time=201.107..201.133 rows=50 loops=1)
          ->  Sort  (cost=31701.98..31989.50 rows=115009 width=1876) (actual time=201.106..201.127 rows=50 loops=1)
                Sort Key: __local_0__.id
                Sort Method: top-N heapsort  Memory: 124kB
                ->  Seq Scan on articles __local_0__  (cost=0.00..26731.38 rows=115009 width=1876) (actual time=0.353..110.675 rows=110587 loops=1)
                      Filter: (status = 'PUBLISHED'::status)
                      Rows Removed by Filter: 6208
Planning time: 4.358 ms
Execution time: 203.685 ms
```

**with `.*`, with CTE**
```
CTE Scan on test  (cost=31702.11..31703.23 rows=50 width=32) (actual time=196.084..197.368 rows=50 loops=1)
  CTE test
    ->  Limit  (cost=31701.98..31702.11 rows=50 width=1930) (actual time=196.015..196.030 rows=50 loops=1)
          ->  Sort  (cost=31701.98..31989.50 rows=115009 width=1930) (actual time=196.013..196.025 rows=50 loops=1)
                Sort Key: __local_0__.id
                Sort Method: top-N heapsort  Memory: 128kB
                ->  Subquery Scan on __local_0__  (cost=0.00..27881.47 rows=115009 width=1930) (actual time=0.313..146.215 rows=110587 loops=1)
                      ->  Seq Scan on articles __local_0___1  (cost=0.00..26731.38 rows=115009 width=1930) (actual time=0.313..138.437 rows=110587 loops=1)
                            Filter: (status = 'PUBLISHED'::status)
                            Rows Removed by Filter: 6208
Planning time: 29.157 ms
Execution time: 197.405 ms
```

So, my suggestion would be with `.*` and with CTE. Should I update the PR accordingly?