@benjie 2017-07-07
I strongly suspect if you use `.*` it'll cause issues with intSetQuery. Thanks 🙏